sesame/requirements.txt

conan>=1.34.1
PyYAML>=5.3.1
colorama>=0.4.4
